// demo-ui/src/app/features/generated/dummy/fool/models/FoolFullView.ts
import { FoolForm } from './FoolForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class FoolFullView {

id: string;
form: FoolForm;
canUpdate: boolean;

/* Specific Code Start */
/* Specific Code End */
}